Add failing test for internal links
As reported in
https://github.com/jeffkreeftmeijer/ox-html-stable-ids.el/issues/18, I
ran into this with links to footnotes when using ox-html-stable-ids.

This patch adds a failing test which hopefully outlines a way to
resolve this issue. Currently, the id and hrefs remain empty, so
they're removed, resulting in broken links.

Instead, it'd be great if we could use the reference we use in the
link as the anchor. However, other options would also be acceptible,
as long as it fixes the link issue and produces stable IDs.