Fall back to :value when no :raw-value exists
The test case for internal links passes if
org-html-stable-ids--extract-id falls back to using the element's
:value property when it has no :raw-value property.

This breaks a different test, though, which relies on the return value
being nil if :raw-value doesn't return anything. It might make sense
to only fall back to :value when the :raw-value key doesn't exist at
all.