Fix internal links (#19)
* Add failing test for internal links

As reported in
https://github.com/jeffkreeftmeijer/ox-html-stable-ids.el/issues/18, I
ran into this with links to footnotes when using ox-html-stable-ids.

This patch adds a failing test which hopefully outlines a way to
resolve this issue. Currently, the id and hrefs remain empty, so
they're removed, resulting in broken links.

Instead, it'd be great if we could use the reference we use in the
link as the anchor. However, other options would also be acceptible,
as long as it fixes the link issue and produces stable IDs.

* Fall back to :value when no :raw-value exists

The test case for internal links passes if
org-html-stable-ids--extract-id falls back to using the element's
:value property when it has no :raw-value property.

This breaks a different test, though, which relies on the return value
being nil if :raw-value doesn't return anything. It might make sense
to only fall back to :value when the :raw-value key doesn't exist at
all.

* Only use :value when it's not deferred

Falling back to the :value property works for inline elements where
the :value is string based, like links. For block-level elements like
example blocks, which have nested elements, those elements get
automatically serialized to a string by org-element-property.

The :value fallback should only work for inline elements, so this
patch switches to using org-element-property-raw, which doesn't
serialize the "deferred" elements. To make sure that doesn't happen,
the new org-html-stable-ids--extract-value function only returns
non-deferred :value properties.

* Manually detangle changes to Org document

* Update description of ID creation, explain extracting values